DDL、Model修正
