+++ ToDoListAPI/ToDoListAPI/Infrastracture/DB/SQLServer/TodolistContext.cs
@@ -1,0 +1,56 @@
+﻿using System;
+using System.Collections.Generic;
+using Microsoft.EntityFrameworkCore;
+using ToDoListAPI.Models;
+
+namespace ToDoListAPI.Infrastracture.DB.SQLServer;
+
+public partial class TodolistContext : DbContext
+{
+    public TodolistContext()
+    {
+    }
+
+    public TodolistContext(DbContextOptions<TodolistContext> options)
+        : base(options)
+    {
+    }
+
+    public virtual DbSet<ToDoList> ToDoLists { get; set; }
+
+    protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
+        => optionsBuilder.UseSqlServer("Name=ConnectionStrings:sqlserver");
+
+    protected override void OnModelCreating(ModelBuilder modelBuilder)
+    {
+        modelBuilder.Entity<ToDoList>(entity =>
+        {
+            entity.HasKey(e => e.Id).HasName("PK__to_do_li__3213E83F1B87421C");
+
+            entity.ToTable("to_do_list");
+
+            entity.Property(e => e.Id).HasColumnName("id");
+            entity.Property(e => e.CreatedAt).HasColumnName("created_at");
+            entity.Property(e => e.Deadline).HasColumnName("deadline");
+            entity.Property(e => e.DeletedAt).HasColumnName("deleted_at");
+            entity.Property(e => e.Detail)
+                .HasColumnType("text")
+                .HasColumnName("detail");
+            entity.Property(e => e.Place)
+                .HasMaxLength(50)
+                .IsUnicode(false)
+                .HasColumnName("place");
+            entity.Property(e => e.Remarks)
+                .HasColumnType("text")
+                .HasColumnName("remarks");
+            entity.Property(e => e.Title)
+                .HasColumnType("text")
+                .HasColumnName("title");
+            entity.Property(e => e.UpdatedAt).HasColumnName("updated_at");
+        });
+
+        OnModelCreatingPartial(modelBuilder);
+    }
+
+    partial void OnModelCreatingPartial(ModelBuilder modelBuilder);
+}
