+++ ToDoListAPI/ToDoListAPI/appsettings.json
@@ -7,6 +7,6 @@
   },
   "AllowedHosts": "*",
   "ConnectionStrings": {
-    "sqlserver": "Data Source=localhost;Initial Catalog=classwork;Persist Security Info=True;Trusted_Connection=yes;TrustServerCertificate=True;User ID=sa;Password=localhost"
+    "sqlserver": "Data Source=localhost;Initial Catalog=todolist;Persist Security Info=True;Trusted_Connection=yes;TrustServerCertificate=True;User ID=sa;Password=localhost"
   }
 }
